% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_dist_al}
\alias{arma_dist_al}
\title{Calculate a distance matrix.}
\usage{
arma_dist_al(X)
}
\arguments{
\item{X}{An n-dimensional matrix of points.}
}
\value{
The distance matrix for the point matrix X.
}
\description{
Computes a distance matrix for the supplied matrix of n-dimensional points
             using Euclidean distance.
}
\examples{
# Distance matrix for the two 3D points.
spbalDAS::arma_dist_al(X = rbind(c(1, 2, 3), c(4, 3, 2)))
# Distance matrix for three 2D points.
spbalDAS::arma_dist_al(X = matrix(c(2, 3, 0, 9, 4, 5), nrow = 3, byrow = TRUE))

}
\author{
Phil Davies.
}
