% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAS.R
\name{DAS}
\alias{DAS}
\title{Dynamic Assignment Sampling (DAS).}
\usage{
DAS(pop, n, verbose = FALSE)
}
\arguments{
\item{pop}{Population points.}

\item{n}{Sample size.}

\item{verbose}{Boolean if you want to see any output printed to screen. Helpful if taking a
long time. Default is FALSE i.e. no informational messages are displayed.}
}
\value{
A list containing one variable, \code{$SampleMatrix} Jn by n matrix containing
 population indices, where each row is a candidate sample.
}
\description{
DAS
}
\examples{
# DAS sample ----------------------------------------------

my_matrix <- matrix(sample(0:500, 10000, replace = TRUE), nrow = 100, ncol = 100)
newpop <- matrix(my_matrix, ncol = 2, byrow = TRUE)
n_samples <- 50
sampMx <- spbalDAS::DAS(pop = newpop, n = n_samples, verbose = FALSE)
# display first twenty sample points.
sampMx[1:20]

}
\author{
This function was first written by Blair Robertson in Matlab and later re-written in R by Phil Davies.
}
