% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleDAS.R
\name{doubleDAS_balance_row}
\alias{doubleDAS_balance_row}
\title{MATLAB LinearAssignment balance term for one candidate sample.}
\usage{
doubleDAS_balance_row(Ci, Ai, X)
}
\description{
MATLAB LinearAssignment balance term for one candidate sample.
}
\keyword{internal}
