% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hungarian_solve}
\alias{hungarian_solve}
\title{Wrapper that calls native_lapjv.}
\arguments{
\item{cost}{The cost matrix.}

\item{Assignment}{The assignment vector.}
}
\value{
The total cost.
}
\description{
...
}
\author{
Phil Davies.
}
