% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_lapjv}
\alias{native_lapjv}
\title{Solve the Linear Assignment Problem.}
\usage{
native_lapjv(Cost_R)
}
\arguments{
\item{Cost_R}{The cost matrix.}
}
\value{
A list containing two variables, cost and assignment.
}
\description{
Solves the linear assignment problem using the Kuhn–Munkres
  (Hungarian) algorithm. The previous Jonker–Volgenant port could cycle on
  floating-point DAS costs.
}
\examples{
# Cost matrix for three 3D points. $cost = 1, $assignment = 2, 1, 0
spbalDAS::native_lapjv(Cost_R = rbind(c(1, 2, 0), c(2, 0, 1), c(1, 4, 19)))
# Cost matrix for three 3D points. $cost = 10, $assignment = 1, 0, 2
spbalDAS::native_lapjv(Cost_R = rbind(c(4, 2, 8), c(2, 3, 7), c(3, 1, 6)))

}
\author{
Phil Davies.
}
