% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAS.R
\name{rSBMoranIvec}
\alias{rSBMoranIvec}
\title{rSBMoranIvec.}
\usage{
rSBMoranIvec(C, A, W)
}
\arguments{
\item{C}{Population points.}

\item{A}{Sample size.}

\item{W}{Boolean if you want to see any output printed to screen. Helpful if taking a
long time. Default is FALSE i.e. no informational messages are displayed.}
}
\value{
A list containing one variable, \code{$SampleMatrix} Jn by n matrix containing
 population indices, where each row is a candidate sample.
}
\description{
rSBMoranIvec
}
\author{
This function was first written by Blair Robertson in Matlab and later re-written in R by Phil Davies.
}
