% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinearAssignmentProblem_cpp}
\alias{LinearAssignmentProblem_cpp}
\title{Solve the Linear Assignment Problem.}
\usage{
LinearAssignmentProblem_cpp(
  W_R,
  target_n,
  N,
  initial_J,
  n_threads = 1L,
  verbose = FALSE
)
}
\arguments{
\item{W_R}{A weight matrix.}

\item{target_n}{The number of sample points required.}

\item{N}{The number of points in the population.}

\item{initial_J}{...}

\item{n_threads}{The maximum number of threads to use. Default value is 1.}

\item{verbose}{When set true will display debug and informational messages.}
}
\value{
The sample matrix.
}
\description{
Solves the linear assignment problem using the hungarian algorithm.
}
\examples{
# Distance matrix for the two 3D points.
spbalDAS::arma_dist_al(X = rbind(c(1, 2, 3), c(4, 3, 2)))
# Distance matrix for three 2D points.
spbalDAS::arma_dist_al(X = matrix(c(2, 3, 0, 9, 4, 5), nrow = 3, byrow = TRUE))

}
\author{
Phil Davies.
}
