% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAS.R
\name{rSBMoranIvec}
\alias{rSBMoranIvec}
\title{rSBMoranIvec.}
\usage{
rSBMoranIvec(C, A, W)
}
\arguments{
\item{C}{bla.}

\item{A}{bla.}

\item{W}{bla.}
}
\value{
To be filled in.
}
\description{
Moran’s I is a measure of spatial autocorrelation–how related the values of a variable are based
on the locations where they were measured.
To calculate Moran’s I, we will need to generate a matrix of inverse distance weights.
In the matrix, entries for pairs of points that are close together are higher than for pairs of points that are far apart.
For simplicity, we will treat the latitude and longitude as values on a plane rather than on a sphere–our locations are
close together and far from the poles. When using latitude and longitude coordinates from more distant locations,
it’s wise to calculate distances based on spherical coordinates (the geosphere package can be used).
}
\author{
This function was first written by Blair Robertson in Matlab and later re-written in R by Phil Davies.
}
